+++ redis-om-spring/src/main/java/com/redis/om/spring/repository/query/RedisEnhancedQuery.java
@@ -298,7 +298,7 @@ public class RedisEnhancedQuery implements RepositoryQuery {
       if (maybePageable.isPresent()) {
         Pageable pageable = maybePageable.get();
         if (!pageable.isUnpaged()) {
-          query.limit(Long.valueOf(pageable.getOffset()).intValue(), pageable.getPageSize());
+          query.limit(Math.toIntExact(pageable.getOffset()), pageable.getPageSize());
   
           if (pageable.getSort() != null) {
             for (Order order : pageable.getSort()) {
